refactor(alerts): extract notification transport into notifications.ts
template.ts mixed Handlebars templating with the HTTP transport for
Slack/generic/incident.io webhooks. Move the transport (notifyChannel,
handleSendSlackWebhook, handleSendGenericWebhook, sendGenericWebhook,
delivery metrics) into tasks/checkAlerts/notifications.ts unchanged, so
the upcoming multi-channel dispatch work lands in a focused module.
